Lib

Documentation

type ApiEffect a = IO (Either ApiErr a)

data ApiKey

Constructors

ApiKey

Fields

bot :: ByteString

user :: Maybe ByteString

get :: [Char] -> [Char]

post :: [Char] -> [Char]

(//) :: String -> String -> String

baseAddress :: String

loginAddress :: [Char]

sendAddress :: [Char]

getAddress :: [Char]

botinfoAddress :: [Char]

botlistAddress :: [Char]

makeKey :: ByteString -> Maybe ByteString -> ApiKey

Build an api key from the bot key, and maybe the user key

makeRequest

Arguments

:: FromJSON a

=> String
The url to make the request to, in the format "<request type> <url>"

-> [(ByteString, Maybe ByteString)]
Array of k:v pairs for request parameters

-> ByteString
Auth string to use

-> ApiEffect a
Returned API response, either an error or your response

Make a request to the API

doLogin :: ApiKey -> ApiEffect LoginResponse

doSend :: ApiKey -> ByteString -> ByteString -> ApiEffect SendResponse

doGet :: ApiKey -> [(ByteString, Maybe ByteString)] -> ApiEffect [DataResponse]

doBotList :: ApiKey -> ApiEffect [BotResponse]

doBotInfo :: ApiKey -> Text -> ApiEffect [BotResponse]